Обработчик событий выполнен в ООП стиле
