+++ app/views/admin/tests/_test.html.erb
@@ -1,4 +1,4 @@
-<tr>
+<tr id="edit-test">
   <td class="align-middle"> 
     <%= content_tag :span, test.title, class: 'test-title', data: { test_id: test.id } %>
     <%= render 'form_inline', test: test == @test ? @test : test %>
